refactor(incentive-wages-service): Split incentiveRate routes out of controller
Move route/middleware declarations and response Zod schemas into
incentiveRate.routes.ts; the controller keeps only request-handling logic
via a createIncentiveRateController(service) factory. module.ts now
creates the shared DocumentedRouter and registers routes into it.

No behavior change — verified route-for-route identical OpenAPI doc
before/after, lint clean, all 19 existing tests pass, and a live smoke
check against the running service.
