+++ apps/incentive-wages-service/src/rates/incentiveRate.controller.ts
@@ -1,72 +1,16 @@
-import { extendZodWithOpenApi } from '@asteasolutions/zod-to-openapi';
-import { z } from 'zod';
+import { asyncHandler, notFound, ok } from '../app.module';
 import type { IncentiveRateService } from './incentiveRate.service';
-import { listActiveRateQuerySchema } from './dto/list-active-rate.dto';
-import {
-  asyncHandler,
-  createDocumentedRouter,
-  notFound,
-  ok,
-  requireRoles,
-  trustGatewayIdentity,
-  validate,
-} from '../app.module';
-
-extendZodWithOpenApi(z);
-
-const incentiveRateSchema = z.object({
-  id: z.string().uuid(),
-  rateType: z.enum(['VISIT', 'REFERRAL', 'MEETING', 'TRAINING', 'RETAINER']),
-  referralType: z.enum(['STANDARD', 'ACCOMPANIED']).nullable(),
-  geographyUnitId: z.string().uuid().nullable(),
-  amountInr: z.number().openapi({ example: 100 }),
-  effectiveFrom: z.string().datetime(),
-  effectiveTo: z.string().datetime().nullable(),
-});
-
-const apiErrorSchema = z.object({
-  success: z.literal(false),
-  message: z.string(),
-  errorCode: z.string().openapi({ example: 'VALIDATION_ERROR' }),
-  details: z.record(z.unknown()).optional(),
-});
-
-function envelope<T extends z.ZodTypeAny>(data: T) {
-  return z.object({ success: z.literal(true), message: z.string(), data });
-}
 
 /**
- * Incentive rate HTTP routes. Mounted under the global `api/v1` prefix.
- * Read-only lookup — rates themselves are managed elsewhere (out of scope
- * here); this exists so a caller deciding an ACCOMPANIED_REFERRAL card
- * (FR-SV-4.9) can resolve the amount before creating an incentive event.
+ * Incentive rate request handlers. Mounted under the global `api/v1` prefix
+ * by `incentiveRate.routes.ts`.
  */
-export function createIncentiveRateRouter(service: IncentiveRateService) {
-  const doc = createDocumentedRouter();
-
-  doc.get(
-    '/incentive-rates/active',
-    {
-      summary: 'Resolve the currently effective incentive rate for a rate/referral type',
-      tags: ['Incentives'],
-      query: listActiveRateQuerySchema,
-      responses: {
-        200: { description: 'Active rate', schema: envelope(incentiveRateSchema) },
-        400: { description: 'Validation error', schema: apiErrorSchema },
-        401: { description: 'Unauthenticated', schema: apiErrorSchema },
-        403: { description: 'Caller role not permitted', schema: apiErrorSchema },
-        404: { description: 'No active rate found', schema: apiErrorSchema },
-      },
-    },
-    trustGatewayIdentity,
-    requireRoles('SUPERVISOR', 'MANAGER', 'ADMIN'),
-    validate(listActiveRateQuerySchema, 'query'),
-    asyncHandler(async (req, res) => {
+export function createIncentiveRateController(service: IncentiveRateService) {
+  return {
+    findActive: asyncHandler(async (req, res) => {
       const rate = await service.findActive(req.query as never);
       if (!rate) throw notFound('No active incentive rate found.');
       res.json(ok(rate));
     }),
-  );
-
-  return doc;
+  };
 }
