+++ apps/incentive-wages-service/src/rates/incentiveRate.module.ts
@@ -1,15 +1,17 @@
-import type { DocumentedRouter } from '@armman/service-commons';
+import { createDocumentedRouter, type DocumentedRouter } from '../app.module';
 import type { PrismaService } from '../prisma/prisma.service';
 import { IncentiveRateRepository } from './incentiveRate.repository';
 import { IncentiveRateService } from './incentiveRate.service';
-import { createIncentiveRateRouter } from './incentiveRate.controller';
+import { registerIncentiveRateRoutes } from './incentiveRate.routes';
 
 /**
  * Composition root for the incentive-rates feature: wires repository →
- * service → router.
+ * service → routes.
  */
 export function createIncentiveRateModule(prisma: PrismaService): DocumentedRouter {
   const repository = new IncentiveRateRepository(prisma);
   const service = new IncentiveRateService(repository);
-  return createIncentiveRateRouter(service);
+  const doc = createDocumentedRouter();
+  registerIncentiveRateRoutes(doc, service);
+  return doc;
 }
